refactor: header가 비어 있을 때 에러를 반환하는 로직 추가
