+++ internal/application/handler/utils/utils.go
