+++ internal/application/presenter/group_response.go
